# fix: remove useless js-loader in front of mdx-loader