--- a/packages/docusaurus-plugin-content-pages/src/index.ts
+++ b/packages/docusaurus-plugin-content-pages/src/index.ts
@@ -186,7 +186,7 @@ export default function pluginContentPages(
       );
     },
 
-    configureWebpack(config, isServer, {getJSLoader}) {
+    configureWebpack() {
       const {
         admonitions,
         rehypePlugins,
@@ -209,7 +209,6 @@ export default function pluginContentPages(
                 // Trailing slash is important, see https://github.com/facebook/docusaurus/pull/3970
                 .map(addTrailingPathSeparator),
               use: [
-                getJSLoader({isServer}),
                 {
                   loader: require.resolve('@docusaurus/mdx-loader'),
                   options: {