## Motivation

It is not needed to put a js-loader (Babel/SWC) on top of the mdx compiler.

The output of MDX is already good to be run directly since MDX v2 and does not need any extra transpilation step (the JSX is now transpiled, unlike in MDX v1 output)

## Site benchmark

I run some basic local tests by building our website locally. 
Unfortunately, the impact of this perf improvement is not as significant as I hoped it to be.

Still, we see some improvement for the Docusaurus site when running on the default Babel JS Loader. 
For SWC users, the impact is even smaller.

Before + Babel
- 92s
- 96s
- 95s
- 97s

Before + SWC
- 87s
- 85s
- 81s
- 86s

After + Babel
- 90s
- 90s
- 91s
- 91s

After + SWC
- 84s
- 85s
- 84s
- 83s

## Test Plan

Preview builds and run as before + unit tests

### Test links


Deploy preview: https://deploy-preview-8972--docusaurus-2.netlify.app/